requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
psycopg2-binary
pydantic-settings
aiofiles
fastapi-cache2[redis]
redis
aioredis
starlette
pydantic
passlib
pydantic_core
sqlmodel
aerich
tortoise-orm
asyncpg
pydantic[email]
bcrypt
argon2_cffi